first stab on live JPSS import and display using geometric computation of fire pixel surface orientation (pixel acquisition time is way too coarse)
added a CesiumLiveApp demo route, including ADS-B tracks, GOES-R and JPSS satellites

renamed race-land into a more fitting race-earth

+++ race-client-ui/src/main/scala/gov/nasa/race/ui/package.scala
@@ -144,7 +144,7 @@ package object ui {
   }
 
   // TODO - suboptimal. In general we want layout&presentation out of the view DSL
-  def uiHorizontalSpacer(minWidthInRem: Int): Text.TypedTag[String] = {
+  def uiHorizontalSpacer(minWidthInRem: Double): Text.TypedTag[String] = {
     val mods = List(cls:="spacer", style:=s"min-width: ${minWidthInRem}rem")
     div(mods: _*)
   }
